feat:componentization of inputs
