+++ future-eats/src/App.js
@@ -1,13 +1,16 @@
-import Rotas from './routes/routes'
-import StyledGlobal from './StyledGlobal';
-import { FutureEatsProvider } from './GlobalContext/FutureEatsProvider';
-
+import Rotas from "./routes/routes";
+import StyledGlobal from "./StyledGlobal";
+import { FutureEatsProvider } from "./GlobalContext/FutureEatsProvider";
+import { ChakraProvider } from "@chakra-ui/react";
+import {theme} from "./styles"
 function App() {
   return (
-    <FutureEatsProvider>
-      <Rotas/>
-      <StyledGlobal/>
-    </FutureEatsProvider>
+    <ChakraProvider theme={theme}>
+      <FutureEatsProvider>
+        <Rotas />
+        <StyledGlobal />
+      </FutureEatsProvider>
+    </ChakraProvider>
   );
 }
