+++ future-eats/src/styles/theme.js
@@ -1,0 +1,5 @@
+import { extendTheme } from '@chakra-ui/react'
+
+export const theme = extendTheme({
+    
+  })
